@@ -2,7 +2,7 @@
     "name": "qs",
     "description": "A querystring parser that supports nesting and arrays, with a depth limit",
     "homepage": "https://github.com/ljharb/qs",
-    "version": "6.14.0",
+    "version": "6.14.1",
     "repository": {
         "type": "git",
         "url": "https://github.com/ljharb/qs.git"
@@ -36,15 +36,15 @@
     "devDependencies": {
         "@browserify/envify": "^6.0.0",
         "@browserify/uglifyify": "^6.0.0",
-        "@ljharb/eslint-config": "^21.1.1",
+        "@ljharb/eslint-config": "^22.1.3",
         "browserify": "^16.5.2",
         "bundle-collapser": "^1.4.0",
         "common-shakeify": "~1.0.0",
         "eclint": "^2.8.1",
-        "es-value-fixtures": "^1.7.0",
-        "eslint": "=8.8.0",
+        "es-value-fixtures": "^1.7.1",
+        "eslint": "^9.39.2",
         "evalmd": "^0.0.19",
-        "for-each": "^0.3.3",
+        "for-each": "^0.3.5",
         "glob": "=10.3.7",
         "has-bigints": "^1.1.0",
         "has-override-mistake": "^1.0.1",
@@ -54,12 +54,13 @@
         "iconv-lite": "^0.5.1",
         "in-publish": "^2.0.1",
         "jackspeak": "=2.1.1",
+        "jiti": "^0.0.0",
         "mkdirp": "^0.5.5",
         "mock-property": "^1.1.0",
         "module-deps": "^6.2.3",
-        "npmignore": "^0.3.1",
+        "npmignore": "^0.3.5",
         "nyc": "^10.3.2",
-        "object-inspect": "^1.13.3",
+        "object-inspect": "^1.13.4",
         "qs-iconv": "^1.0.4",
         "safe-publish-latest": "^2.0.0",
         "safer-buffer": "^2.1.2",
@@ -76,7 +77,7 @@
         "posttest": "npx npm@'>=10.2' audit --production",
         "readme": "evalmd README.md",
         "postlint": "eclint check $(git ls-files | xargs find 2> /dev/null | grep -vE 'node_modules|\\.git' | grep -v dist/)",
-        "lint": "eslint --ext=js,mjs .",
+        "lint": "eslint .",
         "dist": "mkdirp dist && browserify --standalone Qs -g unassertify -g @browserify/envify -g [@browserify/uglifyify --mangle.keep_fnames --compress.keep_fnames --format.indent_level=1 --compress.arrows=false --compress.passes=4 --compress.typeofs=false] -p common-shakeify -p bundle-collapser/plugin lib/index.js > dist/qs.js"
     },
     "license": "BSD-3-Clause",
